Introduce PermissionMonitor

+++ CameraApp/Services/CameraLaunchContextParser.cs
@@ -12,24 +12,24 @@ namespace CameraApp.Services
         {
             var cameraLaunchContext = new CameraLaunchContext();
 
-            var agentId = Parser.ToInt64(args[0]);
-            cameraLaunchContext.AgentId = agentId == 0 ? (long?)null : agentId;
-            cameraLaunchContext.UserId = Parser.ToInt64(args[1]);
+            var agentId = Parser.ToInt32(args[0]);
+            cameraLaunchContext.AgentId = agentId == 0 ? (int?)null : agentId;
+            cameraLaunchContext.UserId = Parser.ToInt32(args[1]);
 
             switch (args.Length)
             {
                 case 4:
                     cameraLaunchContext.StartType = StartType.StartCamera;
-                    cameraLaunchContext.CameraId = Parser.ToInt64(args[2]);
+                    cameraLaunchContext.CameraId = Parser.ToInt32(args[2]);
                     cameraLaunchContext.CameraMode = Parser.ConvertToCameraMode(args[3]);
                     break;
                 case 5:
                     cameraLaunchContext.CameraMode = Parser.ConvertToCameraMode(args[4]);
-                    if (Int64.TryParse(args[2], out var cameraId))
+                    if (Int32.TryParse(args[2], out var cameraId))
                     {
                         cameraLaunchContext.StartType = StartType.StartCameraOnDisplay;
                         cameraLaunchContext.CameraId = cameraId;
-                        cameraLaunchContext.DisplayId = Parser.ToInt64(args[3]);
+                        cameraLaunchContext.DisplayId = Parser.ToInt32(args[3]);
                     }
                     else
                     {
@@ -42,12 +42,12 @@ namespace CameraApp.Services
                     cameraLaunchContext.StartType = StartType.StartVideoSourceOnDisplay;
                     cameraLaunchContext.ServerIp = args[2];
                     cameraLaunchContext.VideoCaptureSource = args[3];
-                    cameraLaunchContext.DisplayId = Parser.ToInt64(args[4]);
+                    cameraLaunchContext.DisplayId = Parser.ToInt32(args[4]);
                     cameraLaunchContext.CameraMode = Parser.ConvertToCameraMode(args[5]);
                     break;
                 case 8:
                     cameraLaunchContext.StartType = StartType.StartCameraInRectangle;
-                    cameraLaunchContext.CameraId = Parser.ToInt64(args[2]);
+                    cameraLaunchContext.CameraId = Parser.ToInt32(args[2]);
                     cameraLaunchContext.Rectangle = new Rectangle
                     {
                         X = Parser.ToInt32(args[3]),
