+++ CameraForms/Dto/CameraLaunchContext.cs
@@ -10,17 +10,17 @@ namespace CameraForms.Dto
 {
     public class CameraLaunchContext
     {
-        public long? AgentId { get; set; }
+        public int? AgentId { get; set; }
 
-        public long UserId { get; set; }
+        public int UserId { get; set; }
 
-        public long CameraId { get; set; }
+        public int CameraId { get; set; }
 
         public string ServerIp { get; set; }
 
         public string VideoCaptureSource { get; set; }
 
-        public long? DisplayId { get; set; }
+        public int? DisplayId { get; set; }
 
         public Rectangle Rectangle { get; set; } = Rectangle.Empty;
